// frontend/src/types/apps/sectionTypes.ts
// Type Imports
import type { ThemeColor } from '@core/types'

export type SectionType = {
    section_id: number
    section_name: string
    dept_id: number
    user_id: number
    email: string
    name: string
    unit_id: number
    unit_name: string
    dept_name: string
    role_id?: string
    company_id?: number
}